Fixed some clang warnings

+++ example/oglplus/016_textured_cube.cpp
@@ -4,7 +4,7 @@
  *
  *  @oglplus_screenshot{016_textured_cube}
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -188,7 +188,7 @@ public:
 		camera_matrix.Set(
 			CamMatrixf::Orbiting(
 				Vec3f(),
-				5.5 - SineWave(time / 6.0) * 3.0,
+				GLfloat(5.5 - SineWave(time / 6.0) * 3.0),
 				FullCircles(time * 0.7),
 				Degrees(SineWave(time / 30.0) * 90)
 			)
