+++ example/oglplus/025_reflected_torus.cpp
@@ -4,7 +4,7 @@
  *
  *  @oglplus_screenshot{025_reflected_torus}
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -307,7 +307,7 @@ public:
 		gl.Viewport(width, height);
 		Mat4f projection = CamMatrixf::PerspectiveX(
 			Degrees(65),
-			double(width)/height,
+			GLfloat(width)/height,
 			1, 40
 		);
 		norm_projection_matrix.Set(projection);
